BulbDialKit: add switch to the next hour at 40 past.
By default, the EMSL code doesn't increment the hour until HH:00. So at
9:59, the hour hand still points at 9. I find this misleading, since on
a real clock the hour hand would be at 10pm.

At 40 minutes past the hour, move the hour hand forward by 1. Now the
hour hand is at 10 from 9:40 to 10:39 instead of 10:00 to 10:59.

I played around with the threshold a bit, and :40 seemed to be the right
time for my "glance-at-the-clock" brain to interpret it correctly. In
particular, :30 seemed to be a bit too early.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>